TDAP061 - Architecture changed, Redux Folder Renamed To Store
