+++ src/index.tsx
@@ -4,15 +4,15 @@ import App from "./App";
 import "bootstrap/dist/css/bootstrap.min.css";
 import "./assets/scss/styles.scss";
 import { Provider } from "react-redux";
-import store from "./redux/store";
+import store from "./store/store";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
 root.render(
   <Provider store={store}>
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
+    <React.StrictMode>
+      <App />
+    </React.StrictMode>
   </Provider>
 );
